Revert the use of ResetNotificationProtocol and EDK2 reset library for variables dump
This reverts e637dc2a059cc02051b7eb52aa414f04c6c37aba and bf473df3aac7c5046f91303e6fdfdbe86a42ad21
The problem is that, as soon as we try to use:
- ArmPkg/Library/ArmSmcPsciResetSystemLib/ArmSmcPsciResetSystemLib.inf
- MdeModulePkg/Universal/ResetSystemRuntimeDxe/ResetSystemRuntimeDxe.inf
instead of:
- ArmPkg/Library/ArmPsciResetSystemLib/ArmPsciResetSystemLib.inf
- EmbeddedPkg/ResetRuntimeDxe/ResetRuntimeDxe.inf
we get a Linux kernel panic on reboot. And this crash occurs even if we don't modify
any code. Just switching to using these modules will make the kernel crash.

As I wasn't able to find a workaround for that issue, I am reverting the
problematic commits so that we can move on with EDK2 integration.

Also apply some minor updates to the Readme's.